+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+
+- Add support and configuration of regex, to ignore files match regex.
+
 ## [0.0.2] - 2020-04-24
 
 ### Changed